fixes based on feedback comments

+++ CSBWebshopSeminarski/CBSWebshopSeminarski.Model/Requests/OrderUpsertRequest.cs
@@ -13,8 +13,8 @@ namespace CBSWebshopSeminarski.Model.Requests
         public DateTime Date { get; set; }
         
         [Required(ErrorMessage = "Price is required.")]
-        [Range(0, float.MaxValue, ErrorMessage = "Price must be zero or greater.")]
-        public float Price { get; set; }
+        [Range(0, double.MaxValue, ErrorMessage = "Price must be zero or greater.")]
+        public decimal Price { get; set; }
         
         [Required(ErrorMessage = "User ID is required.")]
         [Range(1, int.MaxValue, ErrorMessage = "User ID must be valid.")]
